requirements.txt
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
h11==0.16.0
idna==3.10
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
outcome==1.3.0.post0
ply==3.11
PySocks==1.7.1
referencing==0.36.2
requests==2.32.4
robotframework==7.3.2
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.7
robotframework-seleniumlibrary==6.7.1
rpds-py==0.26.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
